+++ libpsx-rs/src/system/gpu/crtc/constants.rs
@@ -1,6 +1,6 @@
-pub(crate) const _CRTC_CLOCK_SPEED_NTSC: f64 = 3.58 * 1e6;
-pub(crate) const _CRTC_CLOCK_SPEED_PAL: f64 = 4.43 * 1e6;
-pub(crate) const REFRESH_RATE_NTSC_PERIOD: f64 = 1.0 / 60.0; // 1 / 60th of a second
-pub(crate) const _REFRESH_RATE_PAL_PERIOD: f64 = 1.0 / 50.0; // 1 / 50th of a second
-pub(crate) const SCANLINE_NTSC_PERIOD: f64 = 63.6e-6; // 63.6 us per scanline
-pub(crate) const _SCANLINE_PAL_PERIOD: f64 = 64.0e-6; // 64.0 us per scanline
+pub(crate) const _CRTC_CLOCK_SPEED_NTSC: f32 = 3.58 * 1e6;
+pub(crate) const _CRTC_CLOCK_SPEED_PAL: f32 = 4.43 * 1e6;
+pub(crate) const REFRESH_RATE_NTSC_PERIOD: f32 = 1.0 / 60.0; // 1 / 60th of a second
+pub(crate) const _REFRESH_RATE_PAL_PERIOD: f32 = 1.0 / 50.0; // 1 / 50th of a second
+pub(crate) const SCANLINE_NTSC_PERIOD: f32 = 63.6e-6; // 63.6 us per scanline
+pub(crate) const _SCANLINE_PAL_PERIOD: f32 = 64.0e-6; // 64.0 us per scanline
