+++ libpsx-rs/src/system/intc/controllers.rs
@@ -17,7 +17,7 @@ pub(crate) fn run(context: &ControllerContext, event: Event) -> ControllerResult
     }
 }
 
-fn run_time(state: &State, duration: f64) -> ControllerResult<()> {
+fn run_time(state: &State, duration: f32) -> ControllerResult<()> {
     let controller_state = &mut state.intc.controller_state.lock();
     controller_state.clock += duration;
