+++ libpsx-rs/src/system/spu/controllers.rs
@@ -30,7 +30,7 @@ pub(crate) fn run(context: &ControllerContext, event: Event) -> ControllerResult
     }
 }
 
-fn run_time(state: &State, audio_backend: &AudioBackend, duration: f64) -> ControllerResult<()> {
+fn run_time(state: &State, audio_backend: &AudioBackend, duration: f32) -> ControllerResult<()> {
     let controller_state = &mut state.spu.controller_state.lock();
     controller_state.clock += duration;
     controller_state.dac_state.clock += duration;
